enhanced-api-phase1/requirements-phase1.txt
"""
Requirements for Phase 1 (Enhanced User API with Database)

Production-ready dependencies with specific versions.
"""

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
email-validator==2.1.0

# Rate Limiting
slowapi==0.1.9

# Caching & Sessions
redis==5.0.1

# Testing & Coverage
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring & Logging
python-logging-loki==0.3.2
prometheus-client==0.19.0

# Utilities
python-multipart==0.0.6
pytz==2023.3

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
